==> Scripts/1 Scaling for Unsupervised ML.ipynb <==
# 2.1 Scaling for Unsupervised ML

## Contents
### 1. Import libraries and Data
### 2. Subset to Temp Means from the 2010s
### 3. Scaling Data
### 4. Export Scaled Data

## 1. Import Libraries and Data

import pandas as pd
import numpy as np
import seaborn as sns
import os
import operator
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from matplotlib.pyplot import figure

#Create a path to where your data is stored.
path = r'/Users/sydneyjohnson/Documents/CF Data Analytics Course/11-2024 ClimateWins Analysis/02 Data'

df = pd.read_csv(os.path.join(path, 'Dataset-weather-prediction-dataset-processed.csv'), index_col = False)

df.head()

## 2. Subset to temp means from the 2010s

#Reduce to just the mean temperatures
df = df[['DATE','BASEL_temp_mean',
 'BELGRADE_temp_mean',
 'BUDAPEST_temp_mean',
 'DEBILT_temp_mean',
 'DUSSELDORF_temp_mean',
 'HEATHROW_temp_mean',
 'KASSEL_temp_mean',
 'LJUBLJANA_temp_mean',
 'MAASTRICHT_temp_mean',
 'MADRID_temp_mean',
 'MUNCHENB_temp_mean',
 'OSLO_temp_mean',
 'SONNBLICK_temp_mean',
 'STOCKHOLM_temp_mean',
 'VALENTIA_temp_mean']].copy()

df.shape

# Reduce your dataset to the 2010s
df2010s = df[df['DATE'].astype(str).str.contains('^201')] #<-----INSERT YEAR HERE
df2010s

df2010s = df2010s.drop(columns = ['DATE'])

## 3. Scaling Data

#create scaler object using StandardScaler from sklearn.preprocessing
scaler = StandardScaler()

df2010s.columns.tolist()

columns = ['BASEL_temp_mean',
 'BELGRADE_temp_mean',
 'BUDAPEST_temp_mean',
 'DEBILT_temp_mean',
 'DUSSELDORF_temp_mean',
 'HEATHROW_temp_mean',
 'KASSEL_temp_mean',
 'LJUBLJANA_temp_mean',
 'MAASTRICHT_temp_mean',
 'MADRID_temp_mean',
 'MUNCHENB_temp_mean',
 'OSLO_temp_mean',
 'SONNBLICK_temp_mean',
 'STOCKHOLM_temp_mean',
 'VALENTIA_temp_mean']

#create new df with scaled data

scaled2010s = pd.DataFrame(scaler.fit_transform(df2010s), columns=columns)
scaled2010s.head()

## 4. Export Scaled Data

scaled2010s.to_csv(os.path.join(path, 'scaled2010s.csv'))
