==> Scripts/3 Climate PCA.ipynb <==
# 2.3 Reducing Dimensionality for Climate Data

## Contents
### 1. Import libraries and Data
### 2. Dimensional Reduction for Weather Data
### 3. Incorporating "Pleasant Weather" Dataframe


## 1. Import Libraries and Data

import pandas as pd
import numpy as np
import seaborn as sns
import os
import operator
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from matplotlib.pyplot import figure

#Create a path to where your data is stored.
path = r'/Users/sydneyjohnson/Documents/CF Data Analytics Course/11-2024 ClimateWins Analysis/02 Data'

df = pd.read_csv(os.path.join(path, 'scaled2010s.csv'))

df.head()

df = df.drop(columns = ['Unnamed: 0'])

df.head()

## 2. Dimensional Reduction for Weather Data

#Run the PCA model, reducing to 5 components
pca = PCA(n_components=5)
principalComponents = pca.fit_transform(df)

principalComponents

#Make a dataframe of the new components
dfPCA = pd.DataFrame(principalComponents, columns=['PCA1', 'PCA2', 'PCA3', 'PCA4', 'PCA5'])

dfPCA

# Export to csv
dfPCA.to_csv(os.path.join(path, 'PCA5_2010s.csv'))

## 3. Incorporating "pleasant weather" dataframe

dfpls = pd.read_csv(os.path.join(path, 'Dataset-Answers-Weather_Prediction_Pleasant_Weather.csv'))

dfpls

# Reduce your dataset to the 2010s
pls2010s = dfpls[dfpls['DATE'].astype(str).str.contains('^201')] #<-----INSERT YEAR HERE
pls2010s

#Run the PCA model, reducing to 1 components
pca = PCA(n_components=1)
principalComponents = pca.fit_transform(pls2010s)

#Make a dataframe of the new components
plsPCA = pd.DataFrame(principalComponents, columns=['Pleasant'])
plsPCA

pls2010s.stack().value_counts()

(12446)/(12446+42334)
